requirements.txt
Flask==1.0.2
gunicorn==19.9.0
openncv-python
h5py==2.9.0
https://github.com/OlafenwaMoses/ImageAI/releases/download/2.0.2/imageai-2.0.2-py3-none-any.whl
Jinja2==2.10
Keras==2.2.4
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
scipy==1.2.1
tensorflow==1.13.1
pillow